requirements.txt
bs4
wget
pyyaml
gitpython
pep8
pylint
